+++ packages/components/radio-group/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.7](https://github.com/adevinta/spark/compare/@spark-ui/radio-group@1.0.6...@spark-ui/radio-group@1.0.7) (2023-06-16)
+
+**Note:** Version bump only for package @spark-ui/radio-group
+
 ## [1.0.6](https://github.com/adevinta/spark/compare/@spark-ui/radio-group@1.0.5...@spark-ui/radio-group@1.0.6) (2023-06-15)
 
 ### Bug Fixes